Fix compiler warnings

+++ libavogadro/src/engines/wireengine.cpp
@@ -172,8 +172,8 @@ namespace Avogadro {
     else if (averageDistance < 5.0)
       width = 2.5;
 
-		// Default to single bond, no stipple
-		short stipple = 0xFFFF;
+    // Default to single bond, no stipple
+    short stipple = static_cast<short>(0xFFFF);
     int order = 1;
     if (m_showMulti) {
       order = b->order();
@@ -182,7 +182,7 @@ namespace Avogadro {
       // For aromatic (dashed bonds)
       if (b->isAromatic()) {
         order = -1;
-				stipple = 0xCCCC;
+	stipple = static_cast<short>(0xCCCC);
       }
     }
